Switch from `boost` to `boost-cpp`
This package only needs a handful of C++ headers and one C++ library
from Boost to build. None of which actually use Python. Nor does this
package use Boost.Python. So it makes sense to switch to the lighter
weight `boost-cpp` package to drop the Python and NumPy parts of Boost
that this does not use.